requirements.txt
google-cloud-texttospeech>=2.34.0
